feat: enabled readonly category in edit tag form
